+++ frontend/mgramseva/lib/router.dart
@@ -5,10 +5,10 @@ import 'package:mgramseva/screeens/Login/Login.dart';
 import 'package:mgramseva/screeens/SearchConnection.dart';
 import 'package:mgramseva/screeens/SelectLanguage/languageSelection.dart';
 import 'package:mgramseva/main.dart';
-import 'package:mgramseva/screeens/Changepassword.dart';
+import 'package:mgramseva/screeens/ChangePassword/Changepassword.dart';
 import 'package:mgramseva/screeens/ConnectionResults.dart';
 import 'package:mgramseva/screeens/Dashboard.dart';
-import 'package:mgramseva/screeens/EditProfile.dart';
+import 'package:mgramseva/screeens/Profile/EditProfile.dart';
 import 'package:mgramseva/screeens/ExpenseDetails.dart';
 import 'package:mgramseva/screeens/GenerateBill/GenerateBill.dart';
 import 'package:mgramseva/screeens/HouseholdDetail.dart';
